requirements.txt
torch~=2.0.1
numpy~=1.22.0
Pillow~=10.0.1
matplotlib~=3.7.3
tqdm~=4.64.1
torchvision~=0.15.2